Integrar cambios del sistema de página principal (#13)

+++ backend/src/user/user.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -6,19 +7,39 @@ import {
   Param,
   Patch,
   Post,
+  UnauthorizedException,
   UseGuards,
 } from '@nestjs/common';
-import { Usuario } from '@prisma/client';
+import { UserTypes } from '@prisma/client';
 import { JwtGuard } from 'src/auth/guard';
 import { User } from './decorator';
 import { UserService } from './user.service';
 import { UserUpdateDto } from './dto/userUpdate.dto';
 import { UserCreateDTO } from './dto/userCreate.dto';
+import { ApiTags } from '@nestjs/swagger';
 
-@UseGuards(JwtGuard) // desactivar para probar
+@ApiTags('Users')
+@UseGuards(JwtGuard)
 @Controller('users')
 export class UserController {
+  private UAE: UnauthorizedException = new UnauthorizedException(
+    'No está autorizado para acceder a este recurso',
+  );
+  private BRE: BadRequestException = new BadRequestException(
+    'El parámetro introducido no tiene el formato correcto',
+  );
+
   constructor(private userService: UserService) {}
+
+  /**
+   * Ruta para obtener a todos los usuarios registrados.
+   * @returns Usuarios encontrados
+   */
+  @Get()
+  getAllUsers(@User('tipoUsuario') userType: UserTypes) {
+    if (userType == UserTypes.Cliente) throw this.UAE;
+    return this.userService.getAllUsers();
+  }
 
   /**
    * Ruta para obtener el usuario logeado.
@@ -26,8 +47,8 @@ export class UserController {
    * @returns Datos del Usuario logeado
    */
   @Get('me')
-  getMe(@User() user: Usuario) {
-    return user;
+  getMe(@User('id') userId: number) {
+    return this.userService.getUserById(userId);
   }
 
   /**
@@ -35,18 +56,11 @@ export class UserController {
    * @param userId ID del usuario
    * @returns Usuario encontrado
    */
-  @Get('get/:userId')
+  @Get(':userId')
   getUserById(@Param('userId') userId: string) {
-    return this.userService.getUserById(+userId);
-  }
+    if (isNaN(parseInt(userId))) throw this.BRE;
 
-  /**
-   * Ruta para obtener a todos los usuarios registrados.
-   * @returns Usuarios encontrados
-   */
-  @Get('get')
-  getAllUsers() {
-    return this.userService.getAllUsers();
+    return this.userService.getUserById(+userId);
   }
 
   /**
@@ -56,9 +70,12 @@ export class UserController {
    * @param dto DTO para crear usuario como administrador
    * @returns Usuario creado
    */
-  @Post('create')
-  createUser(@Body() dto: UserCreateDTO) {
-    //TODO guard de admin
+  @Post()
+  createUser(
+    @User('tipoUsuario') userType: UserTypes,
+    @Body() dto: UserCreateDTO,
+  ) {
+    if (userType == UserTypes.Cliente) throw this.UAE;
     return this.userService.createUser(dto);
   }
 
@@ -69,22 +86,41 @@ export class UserController {
    * @param userId ID del Usuario
    * @returns Usuario borrado
    */
-  @Delete('delete/:userId')
-  deleteUserById(@Param('userId') userId: string) {
-    //TODO guard de admin
+  @Delete(':userId')
+  deleteUserById(
+    @User('tipoUsuario') userType: UserTypes,
+    @Param('userId') userId: string,
+  ) {
+    if (isNaN(parseInt(userId))) throw this.BRE;
+    if (userType == UserTypes.Cliente) throw this.UAE;
     return this.userService.deleteUserById(+userId);
   }
 
   /**
-   * Ruta para actualizar los datos de un usuario dado un ID y el DTO correspondiente
-   *
-   * IMPORTANTE: Se deben incluir todos los datos que NO se modifican. Si no se proporcionan, quedarán en blanco
+   * Ruta para actualizar los datos del usuario loggeado
    * @param userId ID del Usuario
    * @param dto DTO para actualizar el usuario
    * @returns Usuario actualizado
    */
-  @Patch('patch/:userId')
-  updateUserById(@Param('userId') userId: string, @Body() dto: UserUpdateDto) {
+  @Patch()
+  updateLoggedUser(@User('id') userId: number, @Body() dto: UserUpdateDto) {
+    return this.userService.updateUserById(userId, dto);
+  }
+
+  /**
+   * Ruta para actualizar los datos de un usuario dado un ID y el DTO correspondiente
+   * @param userId ID del Usuario
+   * @param dto DTO para actualizar el usuario
+   * @returns Usuario actualizado
+   */
+  @Patch(':userId')
+  updateUserById(
+    @User('tipoUsuario') userType: UserTypes,
+    @Param('userId') userId: string,
+    @Body() dto: UserUpdateDto,
+  ) {
+    if (isNaN(parseInt(userId))) throw this.BRE;
+    if (userType == UserTypes.Cliente) throw this.UAE;
     return this.userService.updateUserById(+userId, dto);
   }
 }
